+++ test/test_group_type.rb
@@ -3,22 +3,11 @@
 require_relative 'helper'
 
 
-# TODO JGrouper::GroupType::TestCase
-class TestGroupType < Test::Unit::TestCase
-
-  def setup
-    @json = { 'json_class' => JGrouper::GroupType,
-              'name'       => 'jgrouper',
-              'uuid'       => 'some-fake-uuid'
-            }.to_json
-    @type = OpenStruct.new :getName => 'jgrouper',
-                           :getUuid => 'some-fake-uuid'
-  end
-
+class JGrouper::GroupType::TestCase
 
   def test_mocked_create_group_type
     Java::EduInternet2MiddlewareGrouper::GrouperSession.expects(:startRootSession).returns true
-    Java::EduInternet2MiddlewareGrouper::GroupType.expects(:createType).returns @type
+    Java::EduInternet2MiddlewareGrouper::GroupType.expects(:createType).returns grouper_type
     JGrouper::GroupType.create('jgrouper') do |group_type|
       assert_not_nil group_type
       assert         group_type.kind_of? JGrouper::GroupType
@@ -29,12 +18,12 @@ class TestGroupType < Test::Unit::TestCase
 
   def test_mocked_group_type_to_json
     Java::EduInternet2MiddlewareGrouper::GrouperSession.expects(:startRootSession).returns true
-    Java::EduInternet2MiddlewareGrouper::GroupType.expects(:createType).returns @type
-    assert_equal @json, JGrouper::GroupType.create('jgrouper').to_json
+    Java::EduInternet2MiddlewareGrouper::GroupType.expects(:createType).returns grouper_type
+    assert_equal jgrouper_type_as_json, JGrouper::GroupType.create('jgrouper').to_json
   end
 
   def test_group_type_from_json
-    group_type = JSON.parse @json
+    group_type = JSON.parse jgrouper_type_as_json
     assert_not_nil group_type
     assert         group_type.kind_of? JGrouper::GroupType
     assert_equal   'jgrouper',       group_type.name
@@ -42,7 +31,7 @@ class TestGroupType < Test::Unit::TestCase
   end
 
   def test_mocked_find_all
-    Java::EduInternet2MiddlewareGrouper::GroupTypeFinder.expects(:findAll).returns [ @type ]
+    Java::EduInternet2MiddlewareGrouper::GroupTypeFinder.expects(:findAll).returns [ grouper_type ]
     group_types = JGrouper::GroupType.all do |group_type|
       assert_not_nil group_type
       assert         group_type.kind_of? JGrouper::GroupType
@@ -57,7 +46,7 @@ class TestGroupType < Test::Unit::TestCase
   end
 
   def test_mocked_find_by_name
-    Java::EduInternet2MiddlewareGrouper::GroupTypeFinder.expects(:find).returns @type
+    Java::EduInternet2MiddlewareGrouper::GroupTypeFinder.expects(:find).returns grouper_type
     JGrouper::GroupType.find('base') do |group_type|
       assert_not_nil  group_type
       assert          group_type.kind_of? JGrouper::GroupType
@@ -72,17 +61,21 @@ class TestGroupType < Test::Unit::TestCase
   end
 
   def test_mocked_instance_delete
+    type = grouper_type
+
     Java::EduInternet2MiddlewareGrouper::GrouperSession.expects(:startRootSession).returns true
-    Java::EduInternet2MiddlewareGrouper::GroupTypeFinder.expects(:find).returns @type
-    @type.expects(:delete).returns(nil)
+    Java::EduInternet2MiddlewareGrouper::GroupTypeFinder.expects(:find).returns type
+    type.expects(:delete).returns(nil)
 
     assert JGrouper::GroupType.find('jgrouper').delete
   end
 
   def test_mocked_instance_delete_failure
+    type = grouper_type
+
     Java::EduInternet2MiddlewareGrouper::GrouperSession.expects(:startRootSession).returns true
-    Java::EduInternet2MiddlewareGrouper::GroupTypeFinder.expects(:find).returns @type
-    @type.expects(:delete).raises(Exception, 'Mocked Exception')
+    Java::EduInternet2MiddlewareGrouper::GroupTypeFinder.expects(:find).returns type
+    type.expects(:delete).raises(Exception, 'Mocked Exception')
 
     group_type = JGrouper::GroupType.find('jgrouper')
     assert_equal false,              group_type.delete
@@ -90,17 +83,21 @@ class TestGroupType < Test::Unit::TestCase
   end
 
   def test_mocked_class_delete
+    type = grouper_type
+
     Java::EduInternet2MiddlewareGrouper::GrouperSession.expects(:startRootSession).returns true
-    Java::EduInternet2MiddlewareGrouper::GroupTypeFinder.expects(:find).returns @type
-    @type.expects(:delete).returns(nil)
+    Java::EduInternet2MiddlewareGrouper::GroupTypeFinder.expects(:find).returns type
+    type.expects(:delete).returns(nil)
 
     assert JGrouper::GroupType.delete('jgrouper')
   end
 
   def test_mocked_class_delete_failure
+    type = grouper_type
+
     Java::EduInternet2MiddlewareGrouper::GrouperSession.expects(:startRootSession).returns true
-    Java::EduInternet2MiddlewareGrouper::GroupTypeFinder.expects(:find).returns @type
-    @type.expects(:delete).raises(Exception, 'Mocked Exception')
+    Java::EduInternet2MiddlewareGrouper::GroupTypeFinder.expects(:find).returns type
+    type.expects(:delete).raises(Exception, 'Mocked Exception')
 
     assert_equal false, JGrouper::GroupType.delete('jgrouper')
   end