+++ test/test_subject.rb
@@ -3,25 +3,10 @@
 require_relative 'helper'
 
 
-# TODO JGrouper::Subject::TestCase
-class TestSubject < Test::Unit::TestCase
-
-  def setup
-    @json = { 'json_class' => JGrouper::Subject,
-              'id'         => 'GrouperSystem',
-              'name'       => 'GrouperSysAdmin',
-              'source'     => 'g:isa',
-              'type'       => 'application'
-            }.to_json
-    @root = OpenStruct.new :getId    => 'GrouperSystem',
-                           :getName  => 'GrouperSysAdmin',
-                           :sourceId => 'g:isa',
-                           :typeName => 'application'
-  end
-
+class JGrouper::Subject::TestCase
 
   def test_mocked_find_root_subject
-    Java::EduInternet2MiddlewareGrouper::SubjectFinder.expects(:findRootSubject).returns @root
+    Java::EduInternet2MiddlewareGrouper::SubjectFinder.expects(:findRootSubject).returns grouper_root_subject
     JGrouper::Subject.root do |subject|
       assert_not_nil subject
       assert         subject.kind_of? JGrouper::Subject
@@ -33,12 +18,12 @@ class TestSubject < Test::Unit::TestCase
   end
 
   def test_mocked_subject_to_json
-    Java::EduInternet2MiddlewareGrouper::SubjectFinder.expects(:findRootSubject).returns @root
-    assert_equal @json, JGrouper::Subject.root.to_json
+    Java::EduInternet2MiddlewareGrouper::SubjectFinder.expects(:findRootSubject).returns grouper_root_subject
+    assert_equal root_subject_as_json, JGrouper::Subject.root.to_json
   end
 
   def test_subject_from_json
-    subject = JSON.parse @json
+    subject = JSON.parse root_subject_as_json
     assert_not_nil subject
     assert         subject.kind_of? JGrouper::Subject
     assert_equal   'GrouperSystem',   subject.id